+++ src/documentRenderers/richtext/prosemirrorPlugins/suggestions/SuggestionPlugin.ts
@@ -1,38 +1,12 @@
 import { Editor, Range } from "@tiptap/core";
 import { escapeRegExp, groupBy } from "lodash";
 import { Plugin, PluginKey, Selection } from "prosemirror-state";
-import { EditorView, Decoration, DecorationSet } from "prosemirror-view";
+import { Decoration, DecorationSet } from "prosemirror-view";
 import SuggestionItem from "./SuggestionItem";
-import createRenderer from "./SuggestionListReactRenderer";
-import { ReplaceStep } from "prosemirror-transform";
-export interface SuggestionRenderer<T extends SuggestionItem> {
-  onExit?: (props: SuggestionRendererProps<T>) => void;
-  onUpdate?: (props: SuggestionRendererProps<T>) => void;
-  onStart?: (props: SuggestionRendererProps<T>) => void;
-  onKeyDown?: (props: SuggestionRendererKeyDownProps) => boolean;
-}
-
-export interface SuggestionRendererKeyDownProps {
-  view: EditorView;
-  event: KeyboardEvent;
-  range: Range;
-}
-
-export type SuggestionRendererProps<T extends SuggestionItem> = {
-  editor: Editor;
-  range: Range;
-  query: string;
-  groups: {
-    [groupName: string]: T[];
-  };
-  count: number;
-  selectItemCallback: (item: T) => void;
-  decorationNode: Element | null;
-  // virtual node for popper.js or tippy.js
-  // this can be used for building popups without a DOM node
-  clientRect: (() => DOMRect) | null;
-  onClose: () => void;
-};
+import createRenderer, {
+  SuggestionRenderer,
+  SuggestionRendererProps,
+} from "./SuggestionListReactRenderer";
 
 export type SuggestionPluginOptions<T extends SuggestionItem> = {
   // Used for ensuring that the plugin key is unique when more than one instance of the SuggestionPlugin is used.
@@ -298,7 +272,7 @@ export function SuggestionPlugin<T extends SuggestionItem>({
           return false;
         }
 
-        return renderer?.onKeyDown?.({ view, event, range }) || false;
+        return renderer?.onKeyDown?.(event) || false;
       },
 
       // Setup decorator on the currently active suggestion.
